gameOver viimeistelty ja game options aloitettu
