--- TasoHyppelyPeli/Assets/Scripts/UI/gameOver.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9929f9a9a33110c43bc519a804a7322c
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
