--- TasoHyppelyPeli/Assets/Scripts/UI/gameOver/gameOver.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 350e02df7e897fb49923bc85d17ba1b9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
