+++ TasoHyppelyPeli/Assets/Scripts/UI/options.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cd5bc2b95ff8f4f4cafa4ea586a46e04
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
