+++ TasoHyppelyPeli/Packages/packages-lock.json
@@ -173,7 +173,7 @@
       "url": "https://packages.unity.com"
     },
     "com.unity.services.core": {
-      "version": "1.5.2",
+      "version": "1.4.3",
       "depth": 1,
       "source": "registry",
       "dependencies": {
