+++ TasoHyppelyPeli/ProjectSettings/EditorBuildSettings.asset
@@ -6,6 +6,9 @@ EditorBuildSettings:
   serializedVersion: 2
   m_Scenes:
   - enabled: 1
+    path: Assets/Scenes/Main Menu.unity
+    guid: 1a96dd7c5dda6cf4d941e7774e2d85b5
+  - enabled: 1
     path: Assets/Scenes/Level 1.unity
     guid: 68ace7028b58e46a7945230722166c92
   - enabled: 1
@@ -14,9 +17,6 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/End Screen.unity
     guid: 7791d6bedbc454d0392ef73db639365a
-  - enabled: 1
-    path: Assets/Scenes/Main Menu.unity
-    guid: 1a96dd7c5dda6cf4d941e7774e2d85b5
   - enabled: 1
     path: Assets/Scenes/gameOver.unity
     guid: 4c95be4475a2b974dbb3be457b13799e
