+++ TasoHyppelyPeli/UserSettings/EditorUserSettings.asset
@@ -18,19 +18,16 @@ EditorUserSettings:
       value: 54060d5701030b5e090c0f7645775b40454e4a2f2a7f75342f7f4e6ab1b66139
       flags: 0
     RecentlyUsedSceneGuid-4:
-      value: 5509550500025909550c597a15770911414e4c7f7e7b76677f7b4c65b4e36d6a
+      value: 52500d5001515e58580a082346205944124e4c7b297f7167792c4f37bab5366d
       flags: 0
     RecentlyUsedSceneGuid-5:
       value: 57520d5307505d0f5a5b0d70127a084412151a792e2d72657a2b4c60b5b96d3d
       flags: 0
-    RecentlyUsedSceneGuid-5:
-      value: 06010405550d5e5d080a5d7749770641434e1a7c7c7d71622b7e446ab1e56560
-      flags: 0
-    RecentlyUsedSceneGuid-5:
-      value: 06010405550d5e5d080a5d7749770641434e1a7c7c7d71622b7e446ab1e56560
-      flags: 0
     RecentlyUsedSceneGuid-6:
-      value: 52500d5001515e58580a082346205944124e4c7b297f7167792c4f37bab5366d
+      value: 5509550500025909550c597a15770911414e4c7f7e7b76677f7b4c65b4e36d6a
+      flags: 0
+    RecentlyUsedSceneGuid-7:
+      value: 06010405550d5e5d080a5d7749770641434e1a7c7c7d71622b7e446ab1e56560
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
